requirements.txt
black==19.10b0
click==8.0.1
lxml==4.6.3
pandas==1.0.1
pyyaml=5.4.1
xlsxwriter==1.2.7
